added header and footer, improved from of track
