+++ package.json
@@ -8,11 +8,16 @@
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
+    "bootstrap-icons": "^1.5.0",
+    "font-awesome": "^4.7.0",
+    "mdbreact": "^5.0.2",
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.0",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-social-icons": "^5.4.0",
     "sass": "^1.32.13",
     "web-vitals": "^1.1.2"
   },
