+++ src/components/TrackList/TrackList.jsx
@@ -24,24 +24,22 @@ componentDidMount(){
 }
 
 render(){
-    const tracks=this.state.data;
-    ;
-    
+    const tracks=this.state.data;    
     return (
         <div class="offset-md-3">
           {tracks.map(track =>
- <div class="card text-center" style={{ width: "48rem" }}>
- <div class="card-body">
-   <h5 class="card-title">{track.name}</h5>
-   <h6 class="card-subtitle text-muted">{track.singer}</h6>
-   <p class="card-text">{track.text}</p>
-   <a href="#" class="btn btn-primary">Go somewhere</a>
- </div>
-</div>)}
+            <div class="card text-center" style={{ width: "48rem" }}>
+                <div class="card-body">
+                <h5 class="card-title">{track.name}</h5>
+                <h6 class="card-subtitle text-muted">{track.singer}</h6>
+                <p class="card-text">{track.text}</p>
+                <a href="#" class="btn btn-secondary">View details</a>
+                </div>
+            </div>)}
         </div>
+
     );
 }
-
 }
 
 export default TrackList;
