--- src/pages/HomePage/HomePage.scss
@@ -1,7 +1,0 @@
-.content{
-    min-height: 80vh;
-}
-
-.сontent-10{
-    min-height: 10vh;
-}
