+++ src/pages/TracksPage/TracksPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import {Footer} from '../../components/footer'
-import {Header} from '../../components/header'
+import {Footer} from '../../components/Footer/footer'
+import {Header} from '../../components/Header/header'
 import {TrackList} from '../../components/TrackList/TrackList'
 
 export class TracksPage extends React.Component{
